improvement(login): rename auth action and reducer to login
